Fix tests for splitText
